Version1 merge

+++ app/src/main/java/com/aoztg/greengrim/data/model/response/CreateNftResponse.kt
@@ -1,5 +1,5 @@
 package com.aoztg.greengrim.data.model.response
 
 data class CreateNftResponse(
-    val nftId: Int
+    val nftId: Long
 )
