+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/challenge/mapper/UiChallengeListMapper.kt
@@ -5,7 +5,7 @@ import com.aoztg.greengrim.presentation.ui.challenge.model.UiChallengeList
 import com.aoztg.greengrim.presentation.ui.challenge.model.UiChallengeRoom
 
 
-fun ChallengeListResponse.toUiChallengeList(onItemClicked: (Int) -> Unit): UiChallengeList{
+fun ChallengeListResponse.toUiChallengeList(onItemClicked: (Long) -> Unit): UiChallengeList{
     return UiChallengeList(
         hasNext = hasNext,
         page = page,
