+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/chat/mapper/UiChatListItemMapper.kt
@@ -4,7 +4,7 @@ import com.aoztg.greengrim.data.model.response.ChatRoomsResponse
 import com.aoztg.greengrim.presentation.ui.chat.model.UiChatListItem
 
 
-fun ChatRoomsResponse.toUiChatListItem(onItemClick: (String, Int, Int) -> Unit): UiChatListItem {
+fun ChatRoomsResponse.toUiChatListItem(onItemClick: (String, Long, Long) -> Unit): UiChatListItem {
     return UiChatListItem(
         chatId = this.chatroomId,
         challengeId = this.challengeId,
