+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/home/mapper/UiNftItemMapper.kt
@@ -3,7 +3,7 @@ package com.aoztg.greengrim.presentation.ui.home.mapper
 import com.aoztg.greengrim.data.model.response.NftListItem
 import com.aoztg.greengrim.presentation.ui.market.model.UiNftItem
 
-internal fun NftListItem.toUiNftItem(onItemClickListener: (Int) -> Unit) = UiNftItem(
+internal fun NftListItem.toUiNftItem(onItemClickListener: (Long) -> Unit) = UiNftItem(
     id = nftAndMemberInfo.nftSimpleInfo.id,
     image = nftAndMemberInfo.nftSimpleInfo.imgUrl,
     title = nftAndMemberInfo.nftSimpleInfo.title,
