+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/home/model/UiHotChallenge.kt
@@ -3,9 +3,9 @@ package com.aoztg.greengrim.presentation.ui.home.model
 import com.aoztg.greengrim.data.model.response.HotChallengeTags
 
 data class UiHotChallenge(
-    val id: Int,
+    val id: Long,
     val imgUrl: String,
     val title: String,
     val keywords: HotChallengeTags,
-    val itemClickListener: (Int) -> Unit
+    val itemClickListener: (Long) -> Unit
 )
