+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/main/MainActivity.kt
@@ -66,9 +66,7 @@ class MainActivity : BaseActivity<ActivityMainBinding>(ActivityMainBinding::infl
         }
     private val cameraPermission = Manifest.permission.CAMERA
     private val notificationPermission = Manifest.permission.POST_NOTIFICATIONS
-
     private lateinit var tempCameraUri: Uri
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         val navHostFragment =
@@ -78,7 +76,7 @@ class MainActivity : BaseActivity<ActivityMainBinding>(ActivityMainBinding::infl
         if (intent.hasExtra("target")) {
             when (intent.getStringExtra("target")) {
                 "GRIM_DETAIL" -> {
-                    intent.getIntExtra("grimId", -1).let {
+                    intent.getLongExtra("grimId", -1).let {
                         navController.toGrimDetail(it)
                     }
                 }
@@ -87,10 +85,9 @@ class MainActivity : BaseActivity<ActivityMainBinding>(ActivityMainBinding::infl
                 }
 
                 "CREATE_NFT" -> {
-                    val grimId = intent.getIntExtra("grimId", -1)
+                    val grimId = intent.getLongExtra("grimId", -1)
                     intent.getStringExtra("grimUrl")?.let {
-                        navController.toCreateNft(grimId, it)
-                    }
+                        navController.toCreateNft(grimId, it) }
                 }
             }
         }
