+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/market/model/UiGrimItem.kt
@@ -1,10 +1,10 @@
 package com.aoztg.greengrim.presentation.ui.market.model
 
 data class UiGrimItem(
-    val id: Int,
+    val id: Long,
     val image: String,
     val title: String,
     val profileImage: String,
     val nickName: String,
-    val navigateToGrimDetail: (Int) -> Unit
+    val navigateToGrimDetail: (Long) -> Unit
 )
